+++ app/src/main/java/edu/wit/mobileapp/mealprepplanner/helper/RecipeIngredient.java
@@ -1,4 +1,4 @@
-package edu.wit.mobileapp.mealprepplanner;
+package edu.wit.mobileapp.mealprepplanner.helper;
 
 public class RecipeIngredient
 {
